Middlewares as Go Plugins PoC
