+++ .gitignore
@@ -19,6 +19,7 @@ testapps/
 /*.conf
 /*.json
 /*.bak
+*.so
 
 /tyk
 logs/
