--- apps/coprocess_app_sample_protected.json
@@ -1,41 +1,0 @@
-{
-    "name": "Tyk Test API (Coprocess auth)",
-    "api_id": "4",
-    "org_id": "default",
-    "definition": {
-        "location": "header",
-        "key": "version"
-    },
-    "auth": {
-        "auth_header_name": "authorization"
-    },
-    "version_data": {
-        "not_versioned": true,
-        "versions": {
-            "Default": {
-                "name": "Default",
-                "expires": "3000-01-02 15:04",
-                "use_extended_paths": true,
-                "extended_paths": {
-                    "ignored": [],
-                    "white_list": [],
-                    "black_list": []
-                }
-            }
-        }
-    },
-    "proxy": {
-        "listen_path": "/coprocess-auth-tyk-api-test/",
-        "target_url": "http://httpbin.org",
-        "strip_listen_path": true
-    },
-    "use_keyless": false,
-    "enable_coprocess_auth": true,
-    "custom_middleware": {
-      "auth_check": {
-        "name": "MyAuthCheck"
-      },
-      "driver": "python"
-    },
-    "enable_batch_request_support": true
-}
