--- apps/coprocess_grpc_app_sample.json
@@ -1,62 +1,0 @@
-{
-    "name": "Tyk Test API (Coprocess + gRPC)",
-    "api_id": "6",
-    "org_id": "default",
-    "definition": {
-        "location": "header",
-        "key": "version"
-    },
-    "auth": {
-        "auth_header_name": "authorization"
-    },
-    "use_keyless": false,
-    "version_data": {
-        "not_versioned": true,
-        "versions": {
-            "Default": {
-                "name": "Default",
-                "expires": "3000-01-02 15:04",
-                "use_extended_paths": true,
-                "extended_paths": {
-                    "ignored": [],
-                    "white_list": [],
-                    "black_list": []
-                }
-            }
-        }
-    },
-    "proxy": {
-        "listen_path": "/grpc/",
-        "target_url": "http://httpbin.org",
-        "strip_listen_path": true
-    },
-    "custom_middleware": {
-      "pre": [
-        {
-          "name": "MyPreMiddleware",
-          "require_session": false
-        }
-      ],
-      "id_extractor": {
-        "extract_from": "header",
-        "extract_with": "value",
-        "extractor_config": {
-          "header_name": "Authorization"
-        }
-      },
-      "driver": "grpc"
-    },
-    "event_handlers": {
-      "events": {
-        "AuthFailure": [
-          {
-            "handler_name": "cp_dynamic_handler",
-            "handler_meta": {
-              "name": "my_handler"
-            }
-          }
-        ]
-      }
-    },
-    "enable_batch_request_support": true
-}
