--- apps/quickstart.json
@@ -1,34 +1,0 @@
-{
-    "name": "Tyk Test API",
-    "api_id": "3",
-    "org_id": "default",
-    "definition": {
-        "location": "",
-        "key": ""
-    },
-    "use_keyless": true,
-    "auth": {
-        "auth_header_name": ""
-    },
-    "version_data": {
-        "not_versioned": true,
-        "versions": {
-            "Default": {
-                "name": "Default",
-                "expires": "3000-01-02 15:04",
-                "use_extended_paths": true,
-                "extended_paths": {
-                    "ignored": [],
-                    "white_list": [],
-                    "black_list": []
-                }
-            }
-        }
-    },
-    "proxy": {
-        "listen_path": "/quickstart/",
-        "target_url": "http://httpbin.org/",
-        "strip_listen_path": true
-    },
-    "do_not_track": true
-}
