+++ gateway/mw_auth_key.go
@@ -1,261 +1,302 @@
 package gateway
 
 import (
-	"errors"
+	"fmt"
 	"net/http"
+	"os"
+	"plugin"
 	"strings"
-	"time"
 
-	"github.com/TykTechnologies/tyk/internal/crypto"
-	"github.com/TykTechnologies/tyk/internal/otel"
-	"github.com/TykTechnologies/tyk/storage"
-
-	"github.com/TykTechnologies/tyk/user"
-
-	"github.com/TykTechnologies/tyk/apidef"
-	"github.com/TykTechnologies/tyk/config"
 	"github.com/TykTechnologies/tyk/request"
-	"github.com/TykTechnologies/tyk/signature_validator"
 )
 
-const (
-	defaultSignatureErrorCode    = http.StatusUnauthorized
-	defaultSignatureErrorMessage = "Request signature verification failed"
-)
+// import (
+// 	"errors"
+// 	"net/http"
+// 	"strings"
+// 	"time"
 
-const (
-	ErrAuthAuthorizationFieldMissing = "auth.auth_field_missing"
-	ErrAuthKeyNotFound               = "auth.key_not_found"
-	ErrAuthCertNotFound              = "auth.cert_not_found"
-	ErrAuthCertExpired               = "auth.cert_expired"
-	ErrAuthKeyIsInvalid              = "auth.key_is_invalid"
+// 	"github.com/TykTechnologies/tyk/internal/crypto"
+// 	"github.com/TykTechnologies/tyk/internal/otel"
+// 	"github.com/TykTechnologies/tyk/storage"
 
-	MsgNonExistentKey  = "Attempted access with non-existent key."
-	MsgNonExistentCert = "Attempted access with non-existent cert."
-	MsgInvalidKey      = "Attempted access with invalid key."
-)
+// 	"github.com/TykTechnologies/tyk/user"
 
-func initAuthKeyErrors() {
-	TykErrors[ErrAuthAuthorizationFieldMissing] = config.TykError{
-		Message: MsgAuthFieldMissing,
-		Code:    http.StatusUnauthorized,
-	}
+// 	"github.com/TykTechnologies/tyk/apidef"
+// 	"github.com/TykTechnologies/tyk/config"
+// 	"github.com/TykTechnologies/tyk/request"
+// 	"github.com/TykTechnologies/tyk/signature_validator"
+// )
 
-	TykErrors[ErrAuthKeyNotFound] = config.TykError{
-		Message: MsgApiAccessDisallowed,
-		Code:    http.StatusForbidden,
-	}
+// const (
+// 	defaultSignatureErrorCode    = http.StatusUnauthorized
+// 	defaultSignatureErrorMessage = "Request signature verification failed"
+// )
 
-	TykErrors[ErrAuthCertNotFound] = config.TykError{
-		Message: MsgApiAccessDisallowed,
-		Code:    http.StatusForbidden,
-	}
+// const (
+// 	ErrAuthAuthorizationFieldMissing = "auth.auth_field_missing"
+// 	ErrAuthKeyNotFound               = "auth.key_not_found"
+// 	ErrAuthCertNotFound              = "auth.cert_not_found"
+// 	ErrAuthCertExpired               = "auth.cert_expired"
+// 	ErrAuthKeyIsInvalid              = "auth.key_is_invalid"
 
-	TykErrors[ErrAuthKeyIsInvalid] = config.TykError{
-		Message: MsgApiAccessDisallowed,
-		Code:    http.StatusForbidden,
-	}
+// 	MsgNonExistentKey  = "Attempted access with non-existent key."
+// 	MsgNonExistentCert = "Attempted access with non-existent cert."
+// 	MsgInvalidKey      = "Attempted access with invalid key."
+// )
 
-	TykErrors[ErrAuthCertExpired] = config.TykError{
-		Message: MsgCertificateExpired,
-		Code:    http.StatusForbidden,
-	}
-}
+// func initAuthKeyErrors() {
+// 	TykErrors[ErrAuthAuthorizationFieldMissing] = config.TykError{
+// 		Message: MsgAuthFieldMissing,
+// 		Code:    http.StatusUnauthorized,
+// 	}
 
-// KeyExists will check if the key being used to access the API is in the request data,
-// and then if the key is in the storage engine
-type AuthKey struct {
-	*BaseMiddleware
-}
+// 	TykErrors[ErrAuthKeyNotFound] = config.TykError{
+// 		Message: MsgApiAccessDisallowed,
+// 		Code:    http.StatusForbidden,
+// 	}
 
-func (k *AuthKey) Name() string {
-	return "AuthKey"
-}
+// 	TykErrors[ErrAuthCertNotFound] = config.TykError{
+// 		Message: MsgApiAccessDisallowed,
+// 		Code:    http.StatusForbidden,
+// 	}
 
-func (k *AuthKey) setContextVars(r *http.Request, token string) {
-	// Flatten claims and add to context
-	if !k.Spec.EnableContextVars {
-		return
-	}
-	if cnt := ctxGetData(r); cnt != nil {
-		// Key data
-		cnt["token"] = token
-		ctxSetData(r, cnt)
-	}
-}
+// 	TykErrors[ErrAuthKeyIsInvalid] = config.TykError{
+// 		Message: MsgApiAccessDisallowed,
+// 		Code:    http.StatusForbidden,
+// 	}
 
-// getAuthType overrides BaseMiddleware.getAuthType.
-func (k *AuthKey) getAuthType() string {
-	return apidef.AuthTokenType
-}
+// 	TykErrors[ErrAuthCertExpired] = config.TykError{
+// 		Message: MsgCertificateExpired,
+// 		Code:    http.StatusForbidden,
+// 	}
+// }
 
-func (k *AuthKey) ProcessRequest(_ http.ResponseWriter, r *http.Request, _ interface{}) (error, int) {
-	if ctxGetRequestStatus(r) == StatusOkAndIgnore {
-		return nil, http.StatusOK
-	}
+// // KeyExists will check if the key being used to access the API is in the request data,
+// // and then if the key is in the storage engine
+// type AuthKey struct {
+// 	*BaseMiddleware
+// }
 
-	// skip auth key check if the request is looped.
-	if ses := ctxGetSession(r); ses != nil && !ctxCheckLimits(r) {
-		return nil, http.StatusOK
-	}
+// func (k *AuthKey) Name() string {
+// 	return "AuthKey"
+// }
 
-	key, authConfig := k.getAuthToken(k.getAuthType(), r)
-	var certHash string
+// func (k *AuthKey) setContextVars(r *http.Request, token string) {
+// 	// Flatten claims and add to context
+// 	if !k.Spec.EnableContextVars {
+// 		return
+// 	}
+// 	if cnt := ctxGetData(r); cnt != nil {
+// 		// Key data
+// 		cnt["token"] = token
+// 		ctxSetData(r, cnt)
+// 	}
+// }
 
-	keyExists := false
-	var session user.SessionState
-	updateSession := false
-	if key != "" {
-		key = stripBearer(key)
-	} else if authConfig.UseCertificate && key == "" && r.TLS != nil && len(r.TLS.PeerCertificates) > 0 {
-		log.Debug("Trying to find key by client certificate")
-		certHash = k.Spec.OrgID + crypto.HexSHA256(r.TLS.PeerCertificates[0].Raw)
-		if time.Now().After(r.TLS.PeerCertificates[0].NotAfter) {
-			return errorAndStatusCode(ErrAuthCertExpired)
-		}
+// // getAuthType overrides BaseMiddleware.getAuthType.
+// func (k *AuthKey) getAuthType() string {
+// 	return apidef.AuthTokenType
+// }
 
-		key = k.Gw.generateToken(k.Spec.OrgID, certHash)
-	} else {
-		k.Logger().Info("Attempted access with malformed header, no auth header found.")
-		return errorAndStatusCode(ErrAuthAuthorizationFieldMissing)
-	}
+// func (k *AuthKey) ProcessRequest(_ http.ResponseWriter, r *http.Request, _ interface{}) (error, int) {
+// 	if ctxGetRequestStatus(r) == StatusOkAndIgnore {
+// 		return nil, http.StatusOK
+// 	}
 
-	session, keyExists = k.CheckSessionAndIdentityForValidKey(key, r)
-	key = session.KeyID
-	if !keyExists {
-		// fallback to search by cert
-		session, keyExists = k.CheckSessionAndIdentityForValidKey(certHash, r)
-		if !keyExists {
-			return k.reportInvalidKey(key, r, MsgNonExistentKey, ErrAuthKeyNotFound)
-		}
-	}
+// 	// skip auth key check if the request is looped.
+// 	if ses := ctxGetSession(r); ses != nil && !ctxCheckLimits(r) {
+// 		return nil, http.StatusOK
+// 	}
 
-	if authConfig.UseCertificate {
-		certLookup := session.Certificate
+// 	key, authConfig := k.getAuthToken(k.getAuthType(), r)
+// 	var certHash string
 
-		if r.TLS != nil && len(r.TLS.PeerCertificates) > 0 {
-			certLookup = certHash
-			if session.Certificate != certHash {
-				session.Certificate = certHash
-				updateSession = true
-			}
-		}
+// 	keyExists := false
+// 	var session user.SessionState
+// 	updateSession := false
+// 	if key != "" {
+// 		key = stripBearer(key)
+// 	} else if authConfig.UseCertificate && key == "" && r.TLS != nil && len(r.TLS.PeerCertificates) > 0 {
+// 		log.Debug("Trying to find key by client certificate")
+// 		certHash = k.Spec.OrgID + crypto.HexSHA256(r.TLS.PeerCertificates[0].Raw)
+// 		if time.Now().After(r.TLS.PeerCertificates[0].NotAfter) {
+// 			return errorAndStatusCode(ErrAuthCertExpired)
+// 		}
 
-		if _, err := k.Gw.CertificateManager.GetRaw(certLookup); err != nil {
-			return k.reportInvalidKey(key, r, MsgNonExistentCert, ErrAuthCertNotFound)
-		}
-	}
+// 		key = k.Gw.generateToken(k.Spec.OrgID, certHash)
+// 	} else {
+// 		k.Logger().Info("Attempted access with malformed header, no auth header found.")
+// 		return errorAndStatusCode(ErrAuthAuthorizationFieldMissing)
+// 	}
 
-	// Set session state on context, we will need it later
-	switch k.Spec.BaseIdentityProvidedBy {
-	case apidef.AuthToken, apidef.UnsetAuth:
-		ctxSetSession(r, &session, updateSession, k.Gw.GetConfig().HashKeys)
-		k.setContextVars(r, key)
-		ctxSetSpanAttributes(r, k.Name(), []otel.SpanAttribute{
-			otel.APIKeyAttribute(key),
-			otel.APIKeyAliasAttribute(session.Alias),
-		}...)
-	}
+// 	session, keyExists = k.CheckSessionAndIdentityForValidKey(key, r)
+// 	key = session.KeyID
+// 	if !keyExists {
+// 		// fallback to search by cert
+// 		session, keyExists = k.CheckSessionAndIdentityForValidKey(certHash, r)
+// 		if !keyExists {
+// 			return k.reportInvalidKey(key, r, MsgNonExistentKey, ErrAuthKeyNotFound)
+// 		}
+// 	}
 
-	// Try using org-key format first:
-	if strings.HasPrefix(key, session.OrgID) {
-		err, statusCode := k.validateSignature(r, key[len(session.OrgID):])
-		if err == nil && statusCode == http.StatusOK {
-			return err, statusCode
-		}
-	}
+// 	if authConfig.UseCertificate {
+// 		certLookup := session.Certificate
 
-	// As a second approach, try to use the internal ID that's part of the B64 JSON key:
-	keyID, err := storage.TokenID(key)
-	if err == nil {
-		err, statusCode := k.validateSignature(r, keyID)
-		if err == nil {
-			return err, statusCode
-		}
-	}
+// 		if r.TLS != nil && len(r.TLS.PeerCertificates) > 0 {
+// 			certLookup = certHash
+// 			if session.Certificate != certHash {
+// 				session.Certificate = certHash
+// 				updateSession = true
+// 			}
+// 		}
 
-	// Last try is to take the key as is:
-	return k.validateSignature(r, key)
-}
+// 		if _, err := k.Gw.CertificateManager.GetRaw(certLookup); err != nil {
+// 			return k.reportInvalidKey(key, r, MsgNonExistentCert, ErrAuthCertNotFound)
+// 		}
+// 	}
 
-func (k *AuthKey) reportInvalidKey(key string, r *http.Request, msg string, errMsg string) (error, int) {
-	k.Logger().WithField("key", k.Gw.obfuscateKey(key)).Info(msg)
+// 	// Set session state on context, we will need it later
+// 	switch k.Spec.BaseIdentityProvidedBy {
+// 	case apidef.AuthToken, apidef.UnsetAuth:
+// 		ctxSetSession(r, &session, updateSession, k.Gw.GetConfig().HashKeys)
+// 		k.setContextVars(r, key)
+// 		ctxSetSpanAttributes(r, k.Name(), []otel.SpanAttribute{
+// 			otel.APIKeyAttribute(key),
+// 			otel.APIKeyAliasAttribute(session.Alias),
+// 		}...)
+// 	}
 
-	// Fire Authfailed Event
-	AuthFailed(k, r, key)
+// 	// Try using org-key format first:
+// 	if strings.HasPrefix(key, session.OrgID) {
+// 		err, statusCode := k.validateSignature(r, key[len(session.OrgID):])
+// 		if err == nil && statusCode == http.StatusOK {
+// 			return err, statusCode
+// 		}
+// 	}
 
-	// Report in health check
-	reportHealthValue(k.Spec, KeyFailure, "1")
+// 	// As a second approach, try to use the internal ID that's part of the B64 JSON key:
+// 	keyID, err := storage.TokenID(key)
+// 	if err == nil {
+// 		err, statusCode := k.validateSignature(r, keyID)
+// 		if err == nil {
+// 			return err, statusCode
+// 		}
+// 	}
 
-	return errorAndStatusCode(errMsg)
-}
+// 	// Last try is to take the key as is:
+// 	return k.validateSignature(r, key)
+// }
 
-func (k *AuthKey) validateSignature(r *http.Request, key string) (error, int) {
+// func (k *AuthKey) reportInvalidKey(key string, r *http.Request, msg string, errMsg string) (error, int) {
+// 	k.Logger().WithField("key", k.Gw.obfuscateKey(key)).Info(msg)
 
-	_, authConfig := k.getAuthToken(k.getAuthType(), r)
-	logger := k.Logger().WithField("key", k.Gw.obfuscateKey(key))
+// 	// Fire Authfailed Event
+// 	AuthFailed(k, r, key)
 
-	if !authConfig.ValidateSignature {
-		return nil, http.StatusOK
-	}
+// 	// Report in health check
+// 	reportHealthValue(k.Spec, KeyFailure, "1")
 
-	errorCode := defaultSignatureErrorCode
-	if authConfig.Signature.ErrorCode != 0 {
-		errorCode = authConfig.Signature.ErrorCode
-	}
+// 	return errorAndStatusCode(errMsg)
+// }
 
-	errorMessage := defaultSignatureErrorMessage
-	if authConfig.Signature.ErrorMessage != "" {
-		errorMessage = authConfig.Signature.ErrorMessage
-	}
+// func (k *AuthKey) validateSignature(r *http.Request, key string) (error, int) {
 
-	validator := signature_validator.SignatureValidator{}
-	if err := validator.Init(authConfig.Signature.Algorithm); err != nil {
-		logger.WithError(err).Info("Invalid signature verification algorithm")
-		return errors.New("internal server error"), http.StatusInternalServerError
-	}
+// 	_, authConfig := k.getAuthToken(k.getAuthType(), r)
+// 	logger := k.Logger().WithField("key", k.Gw.obfuscateKey(key))
 
-	signature := r.Header.Get(authConfig.Signature.Header)
+// 	if !authConfig.ValidateSignature {
+// 		return nil, http.StatusOK
+// 	}
 
-	paramName := authConfig.Signature.ParamName
-	if authConfig.Signature.UseParam || paramName != "" {
-		if paramName == "" {
-			paramName = authConfig.Signature.Header
-		}
+// 	errorCode := defaultSignatureErrorCode
+// 	if authConfig.Signature.ErrorCode != 0 {
+// 		errorCode = authConfig.Signature.ErrorCode
+// 	}
 
-		paramValue := r.URL.Query().Get(paramName)
+// 	errorMessage := defaultSignatureErrorMessage
+// 	if authConfig.Signature.ErrorMessage != "" {
+// 		errorMessage = authConfig.Signature.ErrorMessage
+// 	}
 
-		// Only use the paramValue if it has an actual value
-		if paramValue != "" {
-			signature = paramValue
-		}
-	}
+// 	validator := signature_validator.SignatureValidator{}
+// 	if err := validator.Init(authConfig.Signature.Algorithm); err != nil {
+// 		logger.WithError(err).Info("Invalid signature verification algorithm")
+// 		return errors.New("internal server error"), http.StatusInternalServerError
+// 	}
 
-	if signature == "" {
-		logger.Info("Request signature header not found or empty")
-		return errors.New(errorMessage), errorCode
-	}
+// 	signature := r.Header.Get(authConfig.Signature.Header)
 
-	secret := k.Gw.replaceTykVariables(r, authConfig.Signature.Secret, false)
+// 	paramName := authConfig.Signature.ParamName
+// 	if authConfig.Signature.UseParam || paramName != "" {
+// 		if paramName == "" {
+// 			paramName = authConfig.Signature.Header
+// 		}
 
-	if secret == "" {
-		logger.Info("Request signature secret not found or empty")
-		return errors.New(errorMessage), errorCode
-	}
+// 		paramValue := r.URL.Query().Get(paramName)
 
-	if err := validator.Validate(signature, key, secret, authConfig.Signature.AllowedClockSkew); err != nil {
-		logger.WithError(err).Info("Request signature validation failed")
-		return errors.New(errorMessage), errorCode
-	}
+// 		// Only use the paramValue if it has an actual value
+// 		if paramValue != "" {
+// 			signature = paramValue
+// 		}
+// 	}
 
-	return nil, http.StatusOK
-}
+// 	if signature == "" {
+// 		logger.Info("Request signature header not found or empty")
+// 		return errors.New(errorMessage), errorCode
+// 	}
+
+// 	secret := k.Gw.replaceTykVariables(r, authConfig.Signature.Secret, false)
+
+// 	if secret == "" {
+// 		logger.Info("Request signature secret not found or empty")
+// 		return errors.New(errorMessage), errorCode
+// 	}
+
+// 	if err := validator.Validate(signature, key, secret, authConfig.Signature.AllowedClockSkew); err != nil {
+// 		logger.WithError(err).Info("Request signature validation failed")
+// 		return errors.New(errorMessage), errorCode
+// 	}
+
+// 	return nil, http.StatusOK
+// }
 
 func stripBearer(token string) string {
 	if len(token) > 6 && strings.ToUpper(token[0:7]) == "BEARER " {
 		return token[7:]
 	}
 	return token
+}
+
+func StripBearer(token string) string {
+	return stripBearer(token)
+}
+
+type TykMiddlewareAsPlugin interface {
+	TykMiddleware
+	SetBase(*BaseMiddleware)
+}
+
+func LoadAuthKeyMW() func(base *BaseMiddleware) TykMiddleware {
+	plug, err := plugin.Open("/Users/martin/apps/tyk/plugins/authKey/authKey.so")
+	if err != nil {
+		fmt.Println(err)
+		os.Exit(1)
+	}
+
+	symAuthKey, err := plug.Lookup("GetPlugin")
+	if err != nil {
+		fmt.Println(err)
+		os.Exit(1)
+	}
+
+	AuthKey, ok := symAuthKey.(func(base *BaseMiddleware) TykMiddleware)
+	if !ok {
+		fmt.Println("unexpected type from module symbol")
+		os.Exit(1)
+	}
+
+	return AuthKey
 }
 
 // TODO: move this method to base middleware?
